refactor: rename all codebase use of 'clutter' to 'mewsfeed'
