+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "clutter-dev",
+  "name": "mewsfeed",
   "private": true,
   "version": "0.9.0-alpha",
   "workspaces": [
@@ -9,17 +9,17 @@
   "scripts": {
     "start:launcher": "AGENTS=2 npm run network:launcher",
     "network:launcher": "hc s clean && npm run build:happ && UI_PORT=8888 concurrently \"npm start -w ui\" \"npm run launch:happ\" \"holochain-playground\"",
-    "launch:happ": "echo \"pass\" | RUST_LOG=warn hc launch --piped -n $AGENTS workdir/clutter.happ --ui-port $UI_PORT network mdns",
+    "launch:happ": "echo \"pass\" | RUST_LOG=warn hc launch --piped -n $AGENTS workdir/mewsfeed.happ --ui-port $UI_PORT network mdns",
     "start": "AGENTS=2 npm run network:browser",
     "network:browser": "npm run build:happ && concurrently \"npm run playground\" \"npm run start:agents:browser\"",
     "start:agents:browser": "concurrently-repeat \"npm run start:agent:browser\" $AGENTS",
     "start:agent:browser": "cross-env HC_PORT=$(port) HC_ADMIN_PORT=$(port) concurrently \"npm run start:happ:browser\" \"sleep 5 && npm run start:browser -w ui\"",
     "test": "npm run build:happ && npm t -w tests",
-    "start:happ:browser": "hc s clean && echo pass | RUST_LOG=warn hc s --piped -f=$HC_ADMIN_PORT generate ./workdir/clutter.happ --run=$HC_PORT -a clutter network mdns",
+    "start:happ:browser": "hc s clean && echo pass | RUST_LOG=warn hc s --piped -f=$HC_ADMIN_PORT generate ./workdir/mewsfeed.happ --run=$HC_PORT -a mewsfeed network mdns",
     "package": "npm run build:happ && npm run package:ui && hc web-app pack workdir",
     "package:ui": "npm run build -w ui && cd ui/dist && bestzip ../dist.zip *",
     "build:happ": "npm run build:dnas && hc app pack ./workdir",
-    "build:dnas": "npm run build:zomes && hc dna pack ./dnas/clutter/workdir",
+    "build:dnas": "npm run build:zomes && hc dna pack ./dnas/mewsfeed/workdir",
     "build:zomes": "CARGO_TARGET_DIR=target cargo build --release --target wasm32-unknown-unknown",
     "playground": "run-singleton \"holochain-playground\"",
     "lint": "eslint --fix --ext .ts tests --ext .ts ui/src --ext .vue ui/src"
