+++ tests/src/utils.ts
@@ -3,8 +3,8 @@ import path from "path";
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-export const clutterHapp = {
-  path: path.join(__dirname, "../../workdir/clutter.happ"),
+export const mewsfeedHapp = {
+  path: path.join(__dirname, "../../workdir/mewsfeed.happ"),
 };
 
 export const sleep = (ms: number) =>
