+++ ui/src/pages/MewsFeed.vue
@@ -19,12 +19,12 @@
 <script setup lang="ts">
 import CreateMewField from "@/components/CreateMewField.vue";
 import MewList from "@/components/MewList.vue";
-import { useClutterStore } from "@/stores";
+import { useMewsfeedStore } from "@/stores";
 import { pageHeightCorrection } from "@/utils/page-layout";
 import { ActionHash } from "@holochain/client";
 import { onMounted } from "vue";
 
-const store = useClutterStore();
+const store = useMewsfeedStore();
 onMounted(store.fetchMewsFeed);
 
 const onToggleLickMew = (hash: ActionHash) => store.reloadMew(hash);
