+++ workdir/happ.yaml
@@ -1,14 +1,14 @@
 ---
 manifest_version: "1"
-name: clutter
+name: mewsfeed
 description: ~
 roles:
-  - name: clutter
+  - name: mewsfeed
     provisioning:
       strategy: create
       deferred: false
     dna:
-      bundled: "../dnas/clutter/workdir/clutter.dna"
+      bundled: "../dnas/mewsfeed/workdir/mewsfeed.dna"
       properties: ~
       uuid: ~
       version: ~
